# literaryhub_swagger_file.yaml
openapi: 3.0.0
info:
  title: LiteraryHub AI API
  version: 1.0.0
  description: |
    The LiteraryHub AI API provides a robust suite of tools tailored for publishers and authors. It facilitates the creation of audiobooks from texts, generation of custom book covers, and updates to unpublished manuscripts in Word or PDF formats. Key features include plagiarism detection and identification of restricted topics, empowering users to create and refine content with confidence and compliance. This API serves as a comprehensive resource for automating and enhancing the publication process.
paths:
...
  /v1/detect:
    post:
      summary: Detect Restricted Content
      tags:
        - restricted topic detection
      description: Endpoint for detecting restricted content in a list of extracted texts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                extracted_texts:
                  type: array
                  items:
                    type: object
                    required:
                      - page_number
                      - paragraph_index
                      - text
                    properties:
                      page_number:
                        type: integer
                        description: The page number where the text is found.
                        example: 1
                      paragraph_index:
                        type: integer
                        description: The paragraph index on the page.
                        example: 1
                      text:
                        type: string
                        description: The extracted text content.
                        example: "This is a sample text."
            example:
              extracted_texts:
                - page_number: 1
                  paragraph_index: 1
                  text: "This is a sample text."
                - page_number: 1
                  paragraph_index: 2
                  text: "This is another sample text."

      responses:
        200:
          description: The JSON input with the restricted content detection results.
          content:
            application/json:
              schema:
                type: object
                properties:
                  extracted_texts:
                    type: array
                    items:
                      type: object
                      properties:
                        page_number:
                          type: integer
                        paragraph_index:
                          type: integer
                        text:
                          type: string
                        is_safe:
                          type: boolean
                          description: Flag indicating if the text is safe or restricted.
                          example: true
                example:
                  extracted_texts:
                    - page_number: 1
                      paragraph_index: 1
                      text: "This is a sample text."
                      is_safe: true
                    - page_number: 1
                      paragraph_index: 2
                      text: "This is another sample text."
                      is_safe: true
        400:
          description: No extracted texts provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "No extracted texts provided"
      operationId: detectRestrictedContent
  /v2/detect:
    post:
      summary: Detect Restricted Content
      tags:
        - restricted topic detection
      description: Endpoint for detecting restricted content in a list of extracted texts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                extracted_texts:
                  type: array
                  items:
                    type: object
                    required:
                      - page_number
                      - paragraph_index
                      - text
                    properties:
                      page_number:
                        type: integer
                        description: The page number where the text is found.
                        example: 1
                      paragraph_index:
                        type: integer
                        description: The paragraph index on the page.
                        example: 1
                      text:
                        type: string
                        description: The extracted text content.
                        example: "This is a sample text."
            example:
              extracted_texts:
                - page_number: 1
                  paragraph_index: 1
                  text: "This is a sample text."
                - page_number: 1
                  paragraph_index: 2
                  text: "This is another sample text."

      responses:
        200:
          description: The JSON input with the restricted content detection results.
          content:
            application/json:
              schema:
                type: object
                properties:
                  extracted_texts:
                    type: array
                    items:
                      type: object
                      properties:
                        page_number:
                          type: integer
                        paragraph_index:
                          type: integer
                        text:
                          type: string
                        is_safe:
                          type: boolean
                          description: Flag indicating if the text is safe or restricted.
                          example: true
                example:
                  extracted_texts:
                    - page_number: 1
                      paragraph_index: 1
                      text: "This is a sample text."
                      is_safe: true
                    - page_number: 1
                      paragraph_index: 2
                      text: "This is another sample text."
                      is_safe: true
        400:
          description: No extracted texts provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "No extracted texts provided"
      operationId: detectRestrictedContent
...